Refactor doc generation into one single command.
