+++ cmd/cli-reference/cli-documentation.go
@@ -1,11 +1,14 @@
-package cmd
+package main
 
 import (
 	"bytes"
 	"fmt"
-
 	"github.com/olekukonko/tablewriter"
+	"github.com/pkg/errors"
+	"github.com/redhat-developer/odo/cmd"
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
+	"os"
 )
 
 /*
@@ -86,11 +89,74 @@ func referencePrinter(command *cobra.Command, level int) string {
 		"```sh\n"+command.Example+"\n```",
 		command.Long,
 		tableOutput.String(),
-		"```sh\n"+GenerateCLIStructure()+"\n```",
+		"```sh\n"+fmt.Sprint(commandPrinter(command, 0))+"\n```",
 		commandOutput)
 }
 
+func getFlags(flags *pflag.FlagSet) []string {
+	var f []string
+	flags.VisitAll(func(flag *pflag.Flag) {
+		f = append(f, fmt.Sprintf("--%v", flag.Name))
+	})
+	return f
+}
+
+func flattenFlags(flags []string) string {
+	var flagString string
+	for _, flag := range flags {
+		flagString = flagString + flag + " "
+	}
+	return flagString
+}
+
+func commandPrinter(command *cobra.Command, level int) string {
+	var finalCommand string
+	// add indentation
+	for i := 0; i < level; i++ {
+		finalCommand = finalCommand + "    "
+	}
+	finalCommand = finalCommand +
+		command.Name() +
+		" " +
+		flattenFlags(getFlags(command.NonInheritedFlags())) +
+		": " +
+		command.Short +
+		"\n"
+	for _, subcommand := range command.Commands() {
+		finalCommand = finalCommand + commandPrinter(subcommand, level+1)
+	}
+	return finalCommand
+}
+
 // Generates and returns a markdown-formatted CLI reference page for Odo
-func GenerateCLIReference() string {
-	return referencePrinter(rootCmd, 0)
+func main() {
+	var clidoc = &cobra.Command{
+		Use:   "cli-doc",
+		Short: "Generate CLI reference for Odo",
+
+		Example: `  # Generate a markdown-formatted CLI reference page for Odo
+  cli-doc reference > docs/cli-reference.md
+
+  # Generate the CLI structure
+  cli-doc structure`,
+		Args:      cobra.OnlyValidArgs,
+		ValidArgs: []string{"help", "reference", "structure"},
+
+		Run: func(command *cobra.Command, args []string) {
+			switch args[0] {
+			case "reference":
+				fmt.Print(referencePrinter(cmd.RootCmd(), 0))
+			case "structure":
+				fmt.Print(commandPrinter(cmd.RootCmd(), 0))
+			default:
+				fmt.Print(command.Usage())
+			}
+		},
+	}
+
+	err := clidoc.Execute()
+	if err != nil {
+		fmt.Println(errors.Cause(err))
+		os.Exit(1)
+	}
 }
