--- scripts/cli-reference/generate-cli-reference.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"github.com/redhat-developer/odo/cmd"
-)
-
-func main() {
-	fmt.Print(cmd.GenerateCLIReference())
-}
