--- scripts/cli-structure/generate-cli-structure.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"github.com/redhat-developer/odo/cmd"
-)
-
-func main() {
-	fmt.Print(cmd.GenerateCLIStructure())
-}
